LS_walk and LS_hole was merged into LS_locations.
